+++ project/coffeeShop/src/main/java/edu/mum/coffee/service/ProductService.java
@@ -25,7 +25,7 @@ public class ProductService   {
 		productRepository.delete(product);
 	}
 
-	public Product getProduct(int productId) {
+	public Product getProduct(Long productId) {
 		return  productRepository.findOne(productId);
 	}
 
@@ -47,5 +47,4 @@ public class ProductService   {
 	public List<Product> findByProductType(ProductType productType) {
 		 return productRepository.findByProductType(productType);
 	}
-	
 }
